+++ infection-bootstrap.php
@@ -1,3 +1,0 @@
-<?php
-// This initialisation is not need anymore when compatibility with TYPO3 v11 is removed
-$GLOBALS['TYPO3_CONF_VARS']['FE']['ContentObjects'] ??= [];